Change normalization of association matrix.
The matrix is based on normalized associations <= 1. Because we make it
symmetric by computing W + W.T all values are <= 2 and thus a division
by 2 brings all values back to <= 1. I think, this form of normalization
is better because it doesn't set the strongest association to 1. Only
the word with itself should have an association strength of 1 and that
is why the diagonal gets filled with ones.

This kind of normalization allows an intuitive interpretation of the
association matrix as dot products of the semantic pointers we want to
compute from it.

Note, that it might be better to move this preprocessing to another
place because it is somewhat distributed across different places
(W + W.T vs division by 2).